Fix transaction list state
This is a bug from a few PRs ago. Now that it's just a list, this is a
lot simpler. In fact I think this should be factored back into the app,
it's not necessarily a data model question but a computation from the
data model. (txns + filter state = displayed txns)