Add category data store, fetch, and display
Follows the transactions store pattern to add a category store to the
data model. It would be nice to be able to observe the map directly,
maybe I should refactor to an ObservableValue that's a mutable map?

For now. When the budget is selected, call the Categories API to get the
list of budget categories. Display those in a table below the
transactions.

Fixes #15