Add default split percentage
Adds a default split input field to the main app page. Fetch and save it
in browser cookies. Pass the default split to the split modal, which
uses it to set the split.

There's currently odd behavior where setting a split percent like "15.5"
can result in a number that results in a more precise percent being
generated. I think it's due to how the values propagate through (setting
the percent sets the value, which also sets the percentage based on that
value…).

Another problem for another day… the net result is still correct, the
default split still results in the expected split dollar amount.